Refactor new component
